pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0"]

[project]
authors = [
  {name = "Felix Stark", email = "felix.stark@kfw.de"},
  {name = "Ingmar Mueller", email = "ingmar.mueller@kfw.de"},
  {name = "Mario Macai", email = "mario.macai@accenture.com"},
  {name = "Melissa Schmidt", email = "melissa.schmidt@accenture.com"},
  {name = "Georgia Moldovan", email = "georgia.moldovan@accenture.com"},
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = ["green_energy_api"]
description = "This is an integration for the Home Assistant for Green Energy Api."
keywords = ["green_energy", "kfw", "homeassistant"]
name = "homeassistant_green_energy"
readme = "README.md"
requires-python = ">=3.7"
version = "0.1.0"

[project.optional-dependencies]
dev = ["black", "pylint", "pytest-homeassistant-custom-component", "Flask", "pre-commit"]

[project.urls]
"Bug Tracker" = "https://github.com/openkfw/smartenergy.green_energy/issues"
"Homepage" = "https://github.com/openkfw/smartenergy.green_energy"